Refactoring of xchunked_view
